requirements.txt
Flask>=2.3.3
requests>=2.31.0
lxml>=4.9.2
aiohttp>=3.8.1
beautifulsoup4>=4.11.1
pandas>=1.5.3
umap-learn>=0.5.3
plotly>=5.10.0
matplotlib>=3.6.2
scikit-learn>=1.1.2